--- a/hedera-mirror-importer/src/main/resources/db/migration/v2/V2.5.0__undistribute_balance_tables.sql
+++ b/hedera-mirror-importer/src/main/resources/db/migration/v2/V2.5.0__undistribute_balance_tables.sql
@@ -1,0 +1,2 @@
+select undistribute_table('account_balance');
+select undistribute_table('token_balance');