--- a/hedera-mirror-test/k6/src/rest/test/networkSupplyTimestamp.js
+++ b/hedera-mirror-test/k6/src/rest/test/networkSupplyTimestamp.js
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2024 Hedera Hashgraph, LLC
+ * Copyright (C) 2024 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,22 +16,20 @@
 
 import http from 'k6/http';
 
-import {isValidListResponse, RestTestScenarioBuilder} from '../libex/common.js';
-import {balanceListName} from '../libex/constants.js';
+import {isSuccess, RestTestScenarioBuilder} from '../libex/common.js';
 
-const urlTag = '/balances?timestamp=X';
+const urlTag = '/network/supply';
 
-const getUrl = (testParameters) =>
-  `/balances?timestamp=${testParameters['DEFAULT_BALANCE_TIMESTAMP']}&limit=${testParameters['DEFAULT_LIMIT']}`;
+const getUrl = (testParameters) => `/${urlTag}?timestamp=${testParameters['DEFAULT_BALANCE_TIMESTAMP']}`;
 
 const {options, run, setup} = new RestTestScenarioBuilder()
-  .name('balancesTimestamp') // use unique scenario name among all tests
+  .name('networkSupplyTimestamp') // use unique scenario name among all tests
   .tags({url: urlTag})
   .request((testParameters) => {
     const url = `${testParameters['BASE_URL_PREFIX']}${getUrl(testParameters)}`;
     return http.get(url);
   })
-  .check('Balances with timestamp OK', (r) => isValidListResponse(r, balanceListName))
+  .check('Network supply OK', isSuccess)
   .build();
 
-export {getUrl, options, run, setup};
+export {options, run, setup};