**Description**:

* Add a `balancesPublicKeyTimestamp` k6 test
* Add a `networkSupplyTimestamp` k6 test
* Fix common chart not working when Prometheus is disabled
* Fix k6 not running ramp up when using include filters
* Fix k6 include/exclude filters not working with mixed case test names
* Fix `stackgres.worker.enableMetricsExporter` not taking effect
* Fix `stackgres.worker.enablePostgresUtil` not taking effect
* Fix `/balances?account.publickey&timestamp` not working with local table by moving to a separate public key lookup
* Improve balance API performance 10x by un-distributing balance tables
* Improve `/api/v1/network/supply?timestamp` performance by only querying at most 2 partitions
* Reorganize k6 README and put variables into tables

**Related issue(s)**:

Fixes #9139

**Notes for reviewer**:

Un-distributing `account_balance` took 10 minutes and `token_balance` took 33 minutes in my test cluster that had mainnet balance data. Performance test results:

| Scenario | URL | VUS | Pass% | RPS | Pass RPS | Avg. Req Duration |
|----------|----------|-----|-------|-----|----------|-------------------|
| balancesAccountTimestamp | /balances?account.id=eq:1459478&timestamp=1698342740.152 | 1500 | 100.00 | 6665.52/s | 6665.52/s | 200.22ms |
| balancesPublicKeyTimestamp | /balances?account.publickey=15706...&timestamp=1698342740.152 | 1500 | 99.94 | 3173.93/s | 3172.03/s | 467.91ms | No | |
| balancesTimestamp | /balances?timestamp=1698342740.152&limit=100 | 1500 | 100.00 | 4160.86/s | 4160.86/s | 348.29ms |
| networkSupplyTimestamp | /network/supply?timestamp=1698342740.152 | 1500 | 100.00 | 4043.42/s | 4043.42/s | 363.15ms | No | |
| tokensIdBalancesTimestamp | /tokens/2186153/balances?timestamp=1724814490.78 | 1500 | 100.00 | 5963.29/s | 5963.29/s | 250.30ms |

**Checklist**

- [x] Documented (Code comments, README, etc.)
- [x] Tested (unit, integration, etc.)